Fixes initial guesses bug for corner patches
When trying to normalize POU for corner patch, sometimes the initial guesses grid for inverting the parametrization of a neighboring smooth patch was to coarse, causing nonconvergence. I explicitly enforced that the first inversion must be in the patch (not just in image of parametrization) of neighboring smooth patch and also improved the initial guess.